tools: validatechroot: refactor as a lib instead of main package
