+++ toolkit/tools/pkg/validatechroot/config.go
@@ -1,0 +1,9 @@
+package validatechroot
+
+type Config struct {
+	ToolchainRpmsDir       string
+	TmpDir                 string
+	WorkerTar              string
+	WorkerManifest         string
+	LeaveChrootFilesOnDisk bool
+}
